forced tooltip to flip based on browser width
