+++ src/_scss/pages/agencyV2/visualizations/_totalObligationsOverTime.scss
@@ -60,87 +60,107 @@
                 overflow: visible;
             }
     
-            .budgetary-resources__tooltip-wrapper {
-                height: 100%;
-                width: 100%;
-                .tooltip {
-                    // Don't show the tooltip in mobile
-                    display: none;
-                }
-                >div {
-                    width: 100%;
-                    height: 100%;
-                }    
-                @include media($medium-screen) {
-                    &:hover {
-                        cursor: pointer;
-                    }
-                    .tooltip {
-                        display: block;
-                    }
-                }
-                .tooltip-pointer {
-                    &.left {
-                        left: rem(-16);
-                        @media screen and (-ms-high-contrast: active), (-ms-high-contrast: none) {
-                            // IE only
-                            left: rem(-8);
-                        }
-                    }
-                }
-                .tooltip__content {
-                    width: rem(400);
-                }
-                .budgetary-resources-tooltip {
-                    .tooltip__text {
-                        display: inline-block;
-                        font-weight: $font-semibold;
-                        .budgetary-resources-tooltip__desc {
-                            font-size: rem(16);
-                            float: left;
-                            padding-right: rem(80);
-                        }
-                        .budgetary-resources-tooltip__desc_percent {
-                            font-size: rem(16);
-                            float: right;
-                        }
-                        .budgetary-resources-tooltip__amount {
-                            font-size: rem(22);
-                            float: left;
-                            padding-right: rem(80);
-                        }
-                        .budgetary-resources-tooltip__amount_percent {
-                            font-size: rem(22);
-                            float: right;
-                        }
-                    }
-                }
-                .tooltip__hover-wrapper {
-                    height: 100%;
-                    width: 100%;
-                    @include display(flex);
-                    @include align-items(flex-end);
-                    padding: 0 10%;
-                    @include media($medium-screen) {
-                        padding: 0 rem(2.5);
-                    }
-                }
-                @include media($medium-screen) {
-                    &:hover {
-                        .tooltip__hover-wrapper {
-                            fill: $color-grass-light;
-                            opacity: 0.3;
-                        }
-                    }
-                    &.budgetary-resources__tooltip-wrapper_active {
-                        .tooltip__hover-wrapper {
-                            fill: $color-grass-light;
-                            opacity: 0.1;
-                        }
-                    }
-                }
-            }
-            
+
         }
     }
+
+    .budgetary-resources__tooltip-wrapper {
+        height: 100%;
+        width: 100%;
+        .tooltip {
+            // Don't show the tooltip in mobile
+            display: none;
+        }
+        >div {
+            width: 100%;
+            height: 100%;
+        }
+        @include media($medium-screen) {
+            &:hover {
+                cursor: pointer;
+            }
+            .tooltip {
+                display: block;
+            }
+        }
+        .tooltip-pointer {
+            &.left {
+                left: rem(-16) !important;
+                @media screen and (-ms-high-contrast: active), (-ms-high-contrast: none) {
+                    // IE only
+                    left: rem(-8);
+                }
+            }
+        }
+        .tooltip__content {
+            width: rem(400);
+        }
+
+        @media (max-width: 1400px) {
+            &.tooltip-wrapper .tooltip-spacer {
+                left: -380px !important;
+                z-index: 20 !important;
+            }
+
+            .tooltip-pointer.left {
+                left: 400px !important;
+                transform: rotateY(180deg);
+            }
+        }
+
+        @media (min-width: 1401px) {
+            .tooltip-pointer.bottom {
+                left: 18rem;
+            }
+        }
+        .budgetary-resources-tooltip {
+            .tooltip__text {
+                display: inline-block;
+                font-weight: $font-semibold;
+                .budgetary-resources-tooltip__desc {
+                    font-size: rem(16);
+                    float: left;
+                    padding-right: rem(80);
+                }
+                .budgetary-resources-tooltip__desc_percent {
+                    font-size: rem(16);
+                    float: right;
+                }
+                .budgetary-resources-tooltip__amount {
+                    font-size: rem(22);
+                    float: left;
+                    padding-right: rem(80);
+                }
+                .budgetary-resources-tooltip__amount_percent {
+                    font-size: rem(22);
+                    float: right;
+                }
+            }
+        }
+        .tooltip__hover-wrapper {
+            height: 100%;
+            width: 100%;
+            @include display(flex);
+            @include align-items(flex-end);
+            padding: 0 10%;
+            @include media($medium-screen) {
+                padding: 0 rem(2.5);
+            }
+        }
+        @include media($medium-screen) {
+            &:hover {
+                .tooltip__hover-wrapper {
+                    fill: $color-grass-light;
+                    //opacity: 0.3;
+                }
+            }
+            &.budgetary-resources__tooltip-wrapper_active {
+                .tooltip__hover-wrapper {
+                    fill: $color-grass-light;
+                    //opacity: 0.1;
+                }
+            }
+        }
+    }
+
 }
