+++ src/js/components/agencyV2/visualizations/totalObligationsOverTime/TotalObligationsOverTimeVisualization.jsx
@@ -7,7 +7,9 @@ import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
 import { scaleLinear } from 'd3-scale';
 import { format, getYear } from 'date-fns';
-import { formatNumber } from 'helpers/moneyFormatter';
+import { TooltipWrapper } from "data-transparency-ui";
+
+import { formatMoney, formatNumber, calculatePercentage } from 'helpers/moneyFormatter';
 import { getYDomain, getMilliseconds, determineScenario } from 'helpers/agencyV2/visualizations/TotalObligationsOverTimeVisualizationHelper';
 import { xLabelHeightPlusPadding, yOffsetForPathStrokeWidth, defaultPadding, defaultHeight } from 'dataMapping/agencyV2/visualizations/totalObligationsOverTime';
 import Paths from 'components/agencyV2/visualizations/totalObligationsOverTime/paths/Paths';
@@ -137,63 +139,92 @@ const TotalObligationsOverTimeVisualization = ({
         }
     }, [xDomain, todaysDate]);
 
+    const balance = agencyBudget - data[data.length - 1].obligated;
+    const percentOfTotal = calculatePercentage(agencyBudget - data[data.length - 1].obligated, agencyBudget);
+
+    const tooltip = (
+        <div className="budgetary-resources-tooltip">
+            <div className="tooltip__title">
+                Available Budgetary Resources
+            </div>
+            <div className="tooltip__text">
+                <div className="budgetary-resources-tooltip__desc">Unobligated Balance</div>
+                <div className="budgetary-resources-tooltip__desc_percent">Percent of Total</div>
+                <div className="budgetary-resources-tooltip__amount">{formatMoney(balance)}</div>
+                <div className="budgetary-resources-tooltip__amount_percent">{percentOfTotal}</div>
+            </div>
+        </div>
+    );
+    // {/*// <div className="tooltip-object-overflow" x={rectangleData.x} y={rectangleData.y} width={rectangleData.width + 7} height={rectangleData.height}>*/}
+
     return (
-        <svg
-            className="total-obligations-over-time-svg"
-            height={height}
-            width={width}>
-            <defs>
-                <PathAndAreaPathLinearGradients
-                    agencyBudget={agencyBudget}
-                    data={dataWithFirstAndLastCoordinate}
-                    padding={padding}
-                    width={width} />
-            </defs>
-            <g className="total-obligations-over-time-svg-body">
-                <Paths
-                    data={dataWithFirstAndLastCoordinate}
-                    description={description}
-                    xScale={xScale}
-                    xScaleForPath={xScaleForPath}
-                    yScale={yScale}
-                    yScaleForPath={yScaleForPath}
+        <div className="tooltip-wrapper-overflow">
+            <TooltipWrapper
+                controlledProps={{
+                    isControlled: true,
+                    isVisible: true
+                }}
+                className="budgetary-resources__tooltip-wrapper"
+                offsetAdjustments={{ top: -5 }}
+                tooltipComponent={tooltip}>
+                <svg
+                    className="total-obligations-over-time-svg"
                     height={height}
-                    width={width}
-                    padding={padding}
-                    agencyBudget={agencyBudget}
-                    scenario={scenario} />
-                <Axis
-                    padding={padding}
-                    width={width}
-                    height={height}
-                    xTicks={xTicks} />
-                {showTodayLineAndText && <TodayLineAndtext
-                    xScale={xScale}
-                    height={height}
-                    todaysDate={todaysDate}
-                    padding={padding}
-                    showTodayLineAndText={showTodayLineAndText} />}
-                <AgencyBudgetLine
-                    data={dataWithFirstAndLastCoordinate}
-                    xScale={xScale}
-                    yScale={yScale}
-                    agencyBudget={agencyBudget}
-                    height={height}
-                    width={width}
-                    todaysDate={todaysDate}
-                    padding={padding}
-                    scenario={scenario}
-                    showTodayLineAndText={showTodayLineAndText} />
-                {(scenario === 'exceedsMin' || scenario === 'exceedsMaxAndMin') && <ZeroLineAndTick
-                    xScale={xScale}
-                    yScale={yScale}
-                    height={height}
-                    padding={padding}
-                    width={width}
-                    showTodayLineAndText={showTodayLineAndText}
-                    todaysDate={todaysDate} />}
-            </g>
-        </svg>
+                    width={width}>
+                    <defs>
+                        <PathAndAreaPathLinearGradients
+                            agencyBudget={agencyBudget}
+                            data={dataWithFirstAndLastCoordinate}
+                            padding={padding}
+                            width={width} />
+                    </defs>
+                    <g className="total-obligations-over-time-svg-body">
+                        <Paths
+                            data={dataWithFirstAndLastCoordinate}
+                            description={description}
+                            xScale={xScale}
+                            xScaleForPath={xScaleForPath}
+                            yScale={yScale}
+                            yScaleForPath={yScaleForPath}
+                            height={height}
+                            width={width}
+                            padding={padding}
+                            agencyBudget={agencyBudget}
+                            scenario={scenario} />
+                        <Axis
+                            padding={padding}
+                            width={width}
+                            height={height}
+                            xTicks={xTicks} />
+                        {showTodayLineAndText && <TodayLineAndtext
+                            xScale={xScale}
+                            height={height}
+                            todaysDate={todaysDate}
+                            padding={padding}
+                            showTodayLineAndText={showTodayLineAndText} />}
+                        <AgencyBudgetLine
+                            data={dataWithFirstAndLastCoordinate}
+                            xScale={xScale}
+                            yScale={yScale}
+                            agencyBudget={agencyBudget}
+                            height={height}
+                            width={width}
+                            todaysDate={todaysDate}
+                            padding={padding}
+                            scenario={scenario}
+                            showTodayLineAndText={showTodayLineAndText} />
+                        {(scenario === 'exceedsMin' || scenario === 'exceedsMaxAndMin') && <ZeroLineAndTick
+                            xScale={xScale}
+                            yScale={yScale}
+                            height={height}
+                            padding={padding}
+                            width={width}
+                            showTodayLineAndText={showTodayLineAndText}
+                            todaysDate={todaysDate} />}
+                    </g>
+                </svg>
+            </TooltipWrapper>
+        </div>
     );
 };
